fixed renaming folder
