--- src/TeamControl/Network/__init__.py
@@ -1,18 +1,0 @@
-__all__ = ["BaseSocket","SocketType","Receiver","Broadcast","Multicast","Sender","Broadcaster","Multicaster",
-           "grSimYellowControl","grSimBlueControl","grSimControl","grSimSender","grSimVision","Vision","VisionTracker","GameControl",
-           "GSRobotCommand","RobotCommand"]
-from TeamControl.Network.baseUDP import BaseSocket,SocketType
-from TeamControl.Network.receiver import Receiver, Broadcast, Multicast
-from TeamControl.Network.sender import Sender,Broadcaster,Multicaster
-
-from TeamControl.Network.visionSockets import Vision,VisionTracker
-from TeamControl.Network.grSimSockets import grSimYellowControl,grSimBlueControl,grSimControl,grSimSender,grSimVision
-from TeamControl.Network.gameControllerSockets import GameControl
-
-from TeamControl.Network.robotCommand import RobotCommand
-from TeamControl.Network.grSimRobotCommmand import GSRobotCommand 
-
-if __name__ == "__main__" : 
-    from TeamControl.Network import sender
-    recv = sender()
-
